src/requirements.txt
Flask
requests
pytest
gunicorn
pylint
prometheus_client
mysql-connector-python
